Support updating `Bucket` ACLs (#31)
S3 Buckets support Bucket ACLs in 3 ways:
- Defining a "canned" ACL - through the `ACL` Spec field
- Defining grant headers - through the `Grant*` Spec fields
- Defining a list of grant objects - not supported through `CreateBucket`, and not necessary, so left out

Each of these types is mutually exclusive, and therefore the user is expected to provide at most 1 of the 3. The existing code supports adding ACLs during the `CreateBucket` call, but the `ReadOne` does not provide these values. Instead, to describe these values, we must make a call to [`GetBucketAcl`](https://docs.aws.amazon.com/sdk-for-go/api/service/s3/#S3.GetBucketAcl). 

`GetBucketAcl` returns the ACLs as a list of grant objects. To compare these to the spec fields, we have to write custom code to convert from grant object to spec field type. This comparison is outlined in https://github.com/aws-controllers-k8s/community/issues/895 , and this PR will close this issue. The order of precedence for diffing is canned ACL over `grant` types.

For canned ACLs: A canned ACL is essentially a preset list of grants. Therefore, we can get back to the original canned ACL by comparing the list of grants to the known list of presets and their corresponding settings. However, multiple canned ACLs share the same presets. That is, `private`, `bucket-owner-read` and `bucket-owner-full-control` all return the same grant list. To compare this with `Spec.ACL` (a `string` field), I join the list of possible ACLs into a single string (delimited with `"|"`). In the pre-compare hook, I split the string and compare each of the possibilities to `Spec.ACL`. If any matches, I set that as the `latest.Spec.ACL`, otherwise I set it to `nil`.

For grant headers: There is a many-to-one mapping of grants to grant headers, grouped by `Permission`. We are able to form the grant headers by aggregating by `Permission` and forming the header string by using `ID`, if the `Type` is `CanonicalUser`, or `URI`, if the `Type` is `Group`. Diff works as normal after that.