Fix default canned ACL delta false positive (#80)
Fixes https://github.com/aws-controllers-k8s/community/issues/1282

Description of changes:
The custom ACL code was giving a false positive on the delta when comparing an empty `Spec.ACL` to the list of default ACL possibilities on a fresh bucket. This was causing the controller to always call `SyncACL`, throwing an error if the spec also defined properties which denied ACL use (such as `ObjectOwnership`). This pull request checks for the case where the `Spec.ACL` is nil and where the bucket has the default grants, and will ignore the delta, preventing the false positive.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.