bring controller to rt 0.20.1 and sdk 1.44.93 (#89)
The S3 controller was not successfully being generated with ACK
runtime v0.20.1. An error was being thrown during deepcopy-gen. Turns
out that aws-sdk-go v1.44.93, which was brought in because it is used for
latest code-generator, added a few fields to the S3 Bucket. One of these
fields is `NotificationConfiguration.EventBridgeConfiguration` and it is an
empty struct with no fields. Looks like S3 team just added it into the API
model without finishing the thought... anyway, ignoring this field in the
generator.yaml file enabled the code-generator to successfully complete.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>

Issue aws-controllers-k8s/community#1475

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.