Remove default fields during Adoption
When the controller does a Read operation from AWS, if the field
is empty, it assigns an empty struct to that field. This issue wan't
really a problem, until we introduced the adoption by annotation
feature. When adopting, the controller does a read operation, and
patches every field it finds in the cluster.
Now the assignment of empty structs tends to get in the way, since,
as reported by users, the resource ends up having a bunch of fields
with with empty strings/lists/maps.

The changes introduced here ensures that we only assign a field if it
is not nil, and we don't instanciate an empty field.