Ignore empty fields during Adoption (#150)
Issue [#2230](https://github.com/aws-controllers-k8s/community/issues/2230)

Description of changes:
When the controller does a Read operation from AWS, if the field is empty, 
it assigns an empty struct to that field. This issue wan't really a problem, 
until we introduced the adoption by annotation feature. When adopting, 
the controller does a read operation, and patches every field it finds in the 
cluster.
Now the assignment of empty structs tends to get in the way, since, as 
reported by users, the resource ends up having a bunch of fields with 
empty strings/lists/maps.

The changes introduced here ensures that we only assign a field if it is not nil,
and we don't instanciate an empty field.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.